backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0

# ML & NLP
transformers==4.36.2
torch==2.1.2
peft==0.7.1
sentencepiece==0.1.99
accelerate==0.25.0
safetensors==0.4.1

# Vector Search
faiss-cpu==1.7.4

# Utils
numpy==1.24.3
protobuf==4.25.1